refactor: extract contacts form
